Fixed update of nodeCount in Reduction rules.
The causality graphs are now able to display resources added at different stages.
If after a sequence of reductions, the user adds new resources using the interactive prompt
and starts a new sequence of reductions that use the newly introduced resources, then
these will be displayed in the causality graph.